Fix seed script path, update revenue calculation, and adjust hourly rate type in profile form; enhance API request headers for talent creation

+++ frontend/src/pages/talent/EditProfilePage.jsx
@@ -81,7 +81,7 @@ export default function EditProfilePage() {
         years: parseInt(form.experienceYears) || 0,
         description: form.experienceDesc,
       }));
-      formData.append('hourlyRate', parseFloat(form.hourlyRate) || 0);
+      formData.append('hourlyRate', parseInt(form.hourlyRate) || 0);
       formData.append('location', JSON.stringify({ city: form.city, country: form.country }));
       formData.append('languages', JSON.stringify(form.languages.split(',').map(l => l.trim()).filter(Boolean)));
       formData.append('tags', JSON.stringify(form.tags.split(',').map(t => t.trim()).filter(Boolean)));
